requirements.txt
flask==3.0.0
flask-wtf==1.2.1
python-dotenv==1.0.0
supabase==2.0.3
werkzeug==3.0.1
gunicorn==21.2.0